ASM4 support. Major changes to cglib to support this:
1) *Adapter interfaces have been removed. APIs that implemented *Adapter now extend *Visitor.
    This mostly affects classes that extended other classes and implemented *Adapter, classes
    that implement multiple interfaces, and interfaces that extended *Adapter.  Refactoring
    ensues.
2) ClassWriter no longer allows any of its methods to be overridden, so classes that extended
   it now delegate to it.

Since these changes are made to public classes, the external API has changed. The biggest
changes are probably to ClassTransformer and DebuggingClassWriter.  As a result, we bump the
version number to 3.0.

Also, add words.txt to build process.  Also took the opportunity to change toURL to toURI,
because toURL is deprecated.