Allow to specify a protection domain when defining classes.
At runtime, the CGLIB bundle may be signed by a different authority than
the class to be generated. A typical scenario is with Mockito and OSGi
when generating classes at test runtime for mocking an interface (or a
class). The generated class must use the same protection domain as the
interface it implements. This may be different than CGLIB's.

Task-Url: https://code.google.com/p/mockito/issues/detail?id=393
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>