Fix interpretation of protected and default visibility
There were 2 issues here

VisibilityPredicate would assume that default acceess
(package-private) was always fine as long as the member in question
was in the same package as the generated type.  In fact it needs to
also be in the same classloader too.  This is normally a non-issue since
cglib tries to generate types into the types classloader.  The main
time this doesn't work is for classes on the bootstrap classpath (jdk
types).  So this adds a specific workaround for that.

VisibilityPredicate also assumed that protected access restricted
visibility to subtypes only, this is too strict, protected members are
also visible to types in the same package.

This should ultimately allow guice to drop some !protected checks in a
few places.