Enable reuse of generated by Enhancer classes across GC in case CallbackFilter might become unreachable
In case CallbackFilter becomes unreachable, EnhancerKey's WeakCacheKey<CallbackFilter> field became null,
thus it effectively rendered EnhancerKey to be useless, thus the generated class became inaccessible.

The fix is to add a strong reference from generated class to the CallbackFilter, so the key does not expire,
and the generated class can be reused.

closes #80