Add checkcasts for arguments of invoked bridges
This fixes a TODO in Enhancer: the assumption that a bridge's parameters
are assignable to the target's parameters doesn't hold in certain
cross-compilation scenarios.

The included test case fails with a VerifyError if the checkcasts are
omitted.

Also, make BridgeMethodResolver use the same class loader as Enhancer to
make it easier to test.