Update pom.xml to use the SPDIX license identifier
Update pom.xml to use the SPDIX license indentifier for Apache 2 license.

The benefit of this change is that build automation tools on the market will automatically recognize CGLIB as being an Apache 2.0 licensed library. This is helpful for those organizations that only allow open source libraries licensed under specific whitelisted licenses to be used.